Corrects Remove() functionality
Corrcts previous commit. Added check for node safe for delete in index.
